Mail reminders!!
Uncommented require of action_mailer in application.rb
Added a route to get to the send email code
Added a function in meetups_controller to send an email
reminder_email.html.erb has initial version of the mail to go out
Configured to user my gmail account for sending. In production and dev.
secrets.yml reads the environment variables for the gmail account